Bug fixes in struct and controller

+++ src/models/crypto.go
@@ -1,8 +1,8 @@
 package models
 
 type Crypto struct {
-	ID           int    `json:"id"`
-	Name         string `json:"name"`
-	Amount_Owned int    `json:"amount_owned"`
-	Image_Src    string `json:"image_src"`
+	ID           int     `json:"id"`
+	Name         string  `json:"name"`
+	Amount_Owned float64 `json:"amount_owned"`
+	Image_Src    string  `json:"image_src"`
 }
